requirements.txt
PyGithub>=1.25,<2.0
requests>=2.18,<3.0
just>=0.6,<1.0
scrapy>=1.4,<2.0
